Move relevant tests from shift to alphabet test
