+++ test/shift_test.rb
@@ -39,19 +39,6 @@ class ShiftTest < MiniTest::Test
     assert_equal expected, @enigma.int_from_date(date)
   end
 
-  def test_it_can_create_alphabet
-
-    expected = ["a","b","c","d","e","f","g","h","i","j","k","l","m",
-                "n","o","p","q","r","s","t","u","v","w","x","y","z"," "]
-    assert_equal expected, @enigma.alphabet
-  end
-
-  def test_it_can_shift_alphabet
-    expected = ["c","d","e","f","g","h","i","j","k","l","m","n","o",
-                "p","q","r","s","t","u","v","w","x","y","z"," ","a","b"]
-    assert_equal expected, @enigma.shift_alphabet(2)
-  end
-
   def test_it_creates_random_number_in_correct_range
     (0..10000).each do |i|
       assert @enigma.generate_random_number >= 0
